Update patch set 1
Patch Set 1:

Ron, can you review and submit this at some point this morning.  It makes slight changes to the RTD doc titles so that the index makes sense (as well as building the release notes from the CHANGES), and is the last bit before the bronze branch is created for this repo.   Thanks.

Patch-set: 1