Update patch set 4
Patch Set 4:

Added a small blurb about the RMR section in the xAPP descriptor to the user guide.  Please merge if this is acceptable.    I have a sonar related change that will follow on the heels of this, but don't want to put it in here.   Thanks for the eyes on this.

Patch-set: 4