Add rebuild to unit test script if no lib found
The sonar jenkins jobs do not force a build which is now a problem
for the unit tests as some of them require the framework library
to be found.  This change adds a build step to the unit test
script if the library is missing which will allow the sonar jobs
to run.

Issue-ID: RIC-634

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I34e9a9bf5c6db1a868e390cdfa855e24c0b3b410