Fix $XAPP_DESCRIPTOR_PATH parser bug
While xapp-frame-cpp assumes that $XAPP_DESCRIPTOR_PATH is a filename, the
xapp-onboarder sets it up as a directory. This assumption causes xapps
based on xapp-frame-cpp to crash on startup even using default helm charts.

This change fixes this bug and adds some logic to determine if the
$XAPP_DESCRIPTOR_PATH is a directory or a filename.
This change also adds test cases for $XAPP_DESCRIPTOR_PATH parser.

Issue-ID: RIC-883

Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>
Change-Id: I1d29c8c8ae80028accc5d3e29ea571b6365121ac